chore: move impl for localization to features directory
