+++ app/src/app/[lang]/layout.tsx
@@ -8,9 +8,9 @@ import type { SSRProps } from "@/@types";
 import { Footer } from "@/components/footer";
 import { Header } from "@/components/header";
 import { Providers } from "@/components/provider";
+import { useServerT } from "@/features/localization/hooks/server";
+import { supportedLangs } from "@/features/localization/i18n";
 import { getEnv } from "@/helpers/getEnv";
-import { useServerT } from "@/i18n/server";
-import { supportedLangs } from "@/i18n/settings";
 
 export async function generateMetadata({
   params,
