+++ app/src/app/[lang]/page.tsx
@@ -2,7 +2,7 @@ import type { SSRProps } from "@/@types";
 import { Faq } from "@/components/faq";
 import { Features } from "@/components/features";
 import { Hero } from "@/components/hero";
-import { useServerT } from "@/i18n/server";
+import { useServerT } from "@/features/localization/hooks/server";
 
 export default async function Home({ params }: SSRProps) {
   const { lang } = await params;
