+++ app/src/components/playlists/playlists-root.tsx
@@ -5,9 +5,9 @@ import { useCallback, useEffect, useState } from "react";
 import { type UUID, generateUUID } from "@/actions/generateUUID";
 import { PlaylistManager } from "@/actions/playlist-manager";
 import type { IAdapterPlaylist } from "@/adapters";
+import { useT } from "@/features/localization/hooks/client";
 import { providerToAdapterType } from "@/helpers/providerToAdapterType";
 import { useAuth } from "@/hooks/useAuth";
-import { useT } from "@/i18n/client";
 import { PlaylistActions } from "./playlists-actions";
 import { PlaylistsViewer } from "./playlists-viewer";
 import { type Task, TasksViewer } from "./tasks-viewer";
