+++ app/src/features/localization/resources/en/terms.json
