+++ app/src/features/localization/resources/ja/common.json
