+++ app/src/features/localization/resources/ja/terms.json
